Rename, remove duplicated convert_async_panics
