+++ bindings/core/src/method_handler/mod.rs
@@ -1,7 +1,7 @@
 // Copyright 2023 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-mod account_handle;
+mod account;
 mod call_method;
 mod client;
 mod secret_manager;
